Move method to analysis.py
